make table

+++ database/migrations/2019_01_30_140638_create_company_table.php
@@ -14,8 +14,25 @@ class CreateCompanyTable extends Migration
     public function up()
     {
         Schema::create('company', function (Blueprint $table) {
-            $table->increments('id');
-            $table->timestamps();
+            $table->string('CP_ID', 30);
+            $table->tinyInteger('CP_SEQ');
+            $table->string('CP_NM', 100);
+            $table->string('CP_REG_NO', 15);
+            $table->string('CP_CEO', 50);
+            $table->string('CP_COR_NO', 15);
+            $table->string('CP_TEL', 20);
+            $table->string('CP_FAX', 20);
+            $table->string('CP_ZIP', 15);
+            $table->string('CP_ADD', 200);
+            $table->string('CP_TYPE', 200);
+            $table->string('CP_ITEM', 200);
+            $table->string('CP_BANK', 100);
+            $table->binary('CP_SEAL');
+            $table->string('INPUT_ID', 30);
+            $table->timestamp('INPUT_DATE');
+            $table->string('UPDATE_ID', 30);
+            $table->timestamp('UPDATE_DATE');
+            $table->primary(['CP_ID', 'CP_SEQ']);
         });
     }
 
@@ -26,6 +43,8 @@ class CreateCompanyTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('company');
+        if(Schema::hasTable('company')){
+            Schema::dropIfExists('company');
+        }
     }
 }
