+++ database/migrations/2019_01_30_141025_create_allocation_table.php
@@ -14,8 +14,28 @@ class CreateAllocationTable extends Migration
     public function up()
     {
         Schema::create('allocation', function (Blueprint $table) {
-            $table->increments('id');
-            $table->timestamps();
+            $table->string('CP_ID', 30);
+            $table->string('CA_ID', 30);
+            $table->string('DR_ID', 30);
+            $table->string('AC_ID', 30);
+            $table->string('DS_ID', 30);
+            $table->date('AL_DATE');
+            $table->string('AL_ITEM', 100);
+            $table->tinyInteger('AL_QTY');
+            $table->double('AL_UNIT_COST', 15, 3);
+            $table->string('AL_WORK_TIME', 30);
+            $table->string('AL_LOAD_PLACE', 100);
+            $table->string('AL_DISCHARGE_PLACE', 100);
+            $table->double('AL_DEMAND_AMT', 15, 3);
+            $table->double('AL_PAY_AMT', 15, 3);
+            $table->float('AL_FEE_RATE', 8, 3);
+            $table->double('AL_FEE', 15, 3);
+            $table->string('AL_NOTE', 255);
+            $table->string('INPUT_ID', 30);
+            $table->timestamp('INPUT_DATE');
+            $table->string('UPDATE_ID', 30);
+            $table->timestamp('UPDATE_DATE');
+            $table->primary(['CP_ID', 'CA_ID', 'DR_ID', 'AC_ID', 'DS_ID']);
         });
     }
 
@@ -26,6 +46,9 @@ class CreateAllocationTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('allocation');
+        if(Schema::hasTable('allocation')) {
+            Schema::dropIfExists('allocation');
+        }
+
     }
 }
